requirements.txt
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangotoolbox==1.8.0
djongo==1.3.6
pymongo==3.12.3
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.2.4
tzdata==2022.2
gunicorn==20.1.0
requests==2.28.1
uuid==1.30
dnspython==2.2.1
django-storages==1.13.2
boto3==1.26.64
Pillow==9.4.0
textract-trp==0.1.3
django-crontab==0.7.1
pandas==1.5.3
openpyxl==3.1.1
XlsxWriter==3.0.8
